Add call to project page

+++ database/migrations/2016_03_13_002339_create_price_ranges_table.php
@@ -16,8 +16,8 @@ class CreatePriceRangesTable extends Migration
             $table->increments('id');
             $table->string('key', 250)->unique()->index();
             $table->string('name', 250);
-            $table->integer('from');
-            $table->integer('to');
+            $table->bigInteger('from');
+            $table->bigInteger('to');
             
             $table->integer('priority');
             $table->boolean('active');
